requirements.txt
google-genai
python-dotenv
tenacity